Move fetch article to its own function
